Motor Dance Working!
All console write cmds need to be wrapped in an if statement checking to see if the port is open OTHERWISE THEY FAIL SILENTLY!
Check now occurs in StepperMotor class (which handles all motor movements). If you want to call serial write directly in the future, make sure to check that the port is open!