ready to test slicing & saving
hopefully, slices can be saved as sjfs and we will not have to update the play method. if they cannot be saved as sjfs, we need to think about the next steps for implementation. (next steps will likely involve adding a condition to the play method for the different file type.)